dailybuild: avoid warnings

+++ core/src/Framework/FWTrajectoryPlanner.cpp
@@ -652,7 +652,7 @@ void FWTrajectoryPlanner::HingeJoint::TrajectoryCorrection(int k, bool s) {
 	//終端の到達および停止保証部
 	int stime = (int)(movetime * 0.0);
 
-	for (int i = 0; i < angleLPF.width(); i++) {
+	for (int i = 0; i < (int) angleLPF.width(); i++) {
 		DSTR << k << " " << i << " " << angleLPF[k][i] << std::endl;
 	}
 	double end = angleLPF[k][angleLPF.width() - 1];
@@ -677,7 +677,7 @@ void FWTrajectoryPlanner::HingeJoint::ApplyLPF(int count) {
 		torqueLPF = LPF::centerNSMAv(torqueLPF, 3, 1, initialTorque);
 	}
 	if (count > 0) {
-		for (int i = 0; i < torqueLPF.size(); i++) {
+		for (int i = 0; i < (int) torqueLPF.size(); i++) {
 			torqueLPF[i] = rateLPF * torqueLPF[i] + (1 - rateLPF) * initialTorque;
 		}
 	}
@@ -720,7 +720,7 @@ double FWTrajectoryPlanner::HingeJoint::CalcTotalTorqueChangeLPF() {
 double FWTrajectoryPlanner::HingeJoint::CalcTorqueChangeInSection(int n) {
 	//n = 0, ..., nVia - 1(経由点としては最後), nVia(目標点だけど経由点同様に扱う)
 	double total = 0;
-	if (n >= 0 && n < viatimes.size()) {
+	if (n >= 0 && n < (int) viatimes.size()) {
 		int start = (n == 0) ? 0 : viatimes[n - 1];
 		int end = viatimes[n];
 		for (int i = start; i < end; i++) {
@@ -731,11 +731,11 @@ double FWTrajectoryPlanner::HingeJoint::CalcTorqueChangeInSection(int n) {
 	return total;
 }
 double FWTrajectoryPlanner::HingeJoint::GetBestTorqueChangeInSection(int n) {
-	if (n < 0 || n >= viatimes.size()) return 0;
+	if (n < 0 || n >= (int) viatimes.size()) return 0;
 	return tChanges[n];
 }
 void FWTrajectoryPlanner::HingeJoint::SetBestTorqueChange() {
-	for (int n = 0; n < viatimes.size(); n++) {
+	for (int n = 0; n < (int) viatimes.size(); n++) {
 		tChanges[n] = CalcTorqueChangeInSection(n);
 	}
 }
@@ -956,7 +956,7 @@ void FWTrajectoryPlanner::BallJoint::ApplyLPF(int count) {
 		torqueLPF = LPF::centerNSMAv(torqueLPF, 3, 1, initialTorque);
 	}
 	if (count > 0) {
-		for (int i = 0; i < torqueLPF.size(); i++) {
+		for (int i = 0; i < (int) torqueLPF.size(); i++) {
 			torqueLPF[i] = rateLPF * torqueLPF[i] + (1 - rateLPF) * initialTorque;
 		}
 	}
@@ -1003,7 +1003,7 @@ double FWTrajectoryPlanner::BallJoint::CalcTotalTorqueChangeLPF() {
 }
 double FWTrajectoryPlanner::BallJoint::CalcTorqueChangeInSection(int n) {
 	double total = 0;
-	if (n >= 0 && n < viatimes.size()) {
+	if (n >= 0 && n < (int) viatimes.size()) {
 		int start = (n == 0) ? 0 : viatimes[n - 1];
 		int end = viatimes[n];
 		for (int i = start; i < end; i++) {
@@ -1017,11 +1017,11 @@ double FWTrajectoryPlanner::BallJoint::CalcTorqueChangeInSection(int n) {
 	return total * weight;
 }
 double FWTrajectoryPlanner::BallJoint::GetBestTorqueChangeInSection(int n) {
-	if (n < 0 || n >= viatimes.size()) return 0;
+	if (n < 0 || n >= (int) viatimes.size()) return 0;
 	return tChanges[n];
 }
 void FWTrajectoryPlanner::BallJoint::SetBestTorqueChange() {
-	for(int n = 0; n < viatimes.size(); n++) {
+	for(int n = 0; n < (int) viatimes.size(); n++) {
 		tChanges[n] = CalcTorqueChangeInSection(n);
 	}
 }
@@ -1074,7 +1074,7 @@ FWTrajectoryPlanner::Joints::Joints() { joints.clear(); joints.shrink_to_fit(); 
 FWTrajectoryPlanner::Joints::~Joints() { RemoveAll();  std::vector<Joint*>().swap(joints); }
 void FWTrajectoryPlanner::Joints::RemoveAll() {
 	// 各要素でnewしているためこう？
-	for (int i = 0; i < joints.size(); i++) {
+	for (int i = 0; i < (int) joints.size(); i++) {
 		delete joints[i];
 	}
 	joints.clear();  joints.shrink_to_fit();
@@ -1532,7 +1532,7 @@ void FWTrajectoryPlanner::OutputTrajectory(std::string filename) {
 	// 手先軌道をファイル出力出力
 	std::ofstream outfile(filename + ".csv");
 	for (int i = 0; i < maxIterate + 1; i++) {
-		for (int j = 0; j < trajData.width(); j++) {
+		for (int j = 0; j < (int) trajData.width(); j++) {
 			outfile << trajData[i][j].Pos().x << "," << trajData[i][j].Pos().y << "," << trajData[i][j].Pos().z << "," << std::endl;
 		}
 	}
